requirements/local.txt
-r test.txt

django-debug-toolbar
amqp
asgiref
billiard
braintree
Brotli
celery
certifi
cffi
charset-normalizer
click
click-didyoumean
click-plugins
click-repl
cryptography
cssselect2
defusedxml
django-allauth
django-ckeditor
django-crispy-forms
django-js-asset
django-paypal
django-phonenumber-field
django-taggit
djangorestframework
fonttools
html5lib
idna
kombu
oauthlib
phonenumbers
phonenumberslite
Pillow
prompt-toolkit
psycopg2-binary
pycparser
pydyf
PyJWT
pyphen
python3-openid
pytz
requests
requests-oauthlib
six
sqlparse
tinycss2
tzdata
urllib3
vine
wcwidth
weasyprint
webencodings
whitenoise
zopfli

